requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.2
huggingface-hub==0.32.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pgvector==0.4.1
pillow==11.2.1
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.8.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.3
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
-e git+ssh://git@github.com/jygheo/musinsa-visual-search.git@d2c1a565cf088d76ed27623be917984409e6f224#egg=your_project
